requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
SQLAlchemy>=2.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development tools
ruff>=0.1.0
black>=23.0.0
mypy>=1.6.0

# Optional: For load testing
# locust>=2.16.0

